In order to target iOS 7, I needed to change how I was instantiating SKShapeNode.  I initially utilized a new iOS 8 enhancement to SKShapeNode and instantiated my shape via convenience initializers that are missing in iOS 7, specifically the "rectOfSize" and "cornerRadius" initializing attributes.  In order to make things work under iOS 7, I had to change each instance of SKShapeNode from being instantiated via convenience initializers to being generically initialized and then applying a UIBezierPath to the path of the generic sprite node to facilitate the rounded corners.
Also added new functionality that randomly changes the starting column position from shape to shape.  The only issue is that if a shape is wide.  If the random start column happens to be the first or last columns and the shape has tiles that extend beyond those constraints, the "checkIllegalPlacement" function would incorrectly report that and subsequently end the game.  To circumvent this, I modified the random column start selector algorithm to not include the first or last columns.  Ideally, code would be added to check for this condition before the preview tiles are moved to the playing field and it's starting position adjusted accordingly.